+++ includes/settings.php
@@ -79,6 +79,7 @@ if (!function_exists('tfm_load_settings')) {
             'sitemap_exclude_empty_cats' => true,
             'disable_wp_revisions' => false,
             'wp_post_revisions_limit' => 5,
+            'elementor_acceptance_values' => false,
             'disable_emojis' => false,
             'disable_jquery_migrate' => false,
             'disable_oembeds' => false